+++ src/main/java/com/github/heikyudev/maestrocervecero/presentation/form_dto/insumo/LupuloFormDTO.java
@@ -31,5 +31,5 @@ public class LupuloFormDTO {
     /**
      * Porcentaje de alfa ácidos (AA%). Regla de negocio: positivo, mayor a 0.
      */
-    private Integer aa;
+    private Double aa;
 }
